Move MultiFrameDice to pyro.infer.util
